Rename - Sign up Page Marge(#19)
